third_party/blink/web_tests/fast/css/display-none-documentElement-expected.txt

CONSOLE ERROR: Uncaught TypeError: Failed to execute 'createDocument' on 'DOMImplementation': 2 arguments required, but only 0 present.
